% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbiconf-package.R
\docType{package}
\name{dbiconf-package}
\alias{dbiconf}
\alias{dbiconf-package}
\title{dbiconf: DBI Configuration}
\description{
Specify DBI connection parameters in a
    configuration file. Parameters can be loaded dynamically at run time
    from multiple locations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/shunsambongi/dbiconf}
  \item Report bugs at \url{https://github.com/shunsambongi/dbiconf/issues}
}

}
\author{
\strong{Maintainer}: Shun Sambongi \email{shundev3@gmail.com}

}
\keyword{internal}
