% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loader.R
\name{loader_resolve}
\alias{loader_resolve}
\title{Resolve loader value}
\usage{
loader_resolve(loader)
}
\arguments{
\item{loader}{Loader object (dbiconf_loader) which will be resolved into an
object.}
}
\description{
Resolve loader value
}
