% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_connectors}
\alias{read_connectors}
\title{Read connectors}
\usage{
read_connectors(file)
}
\arguments{
\item{file}{Path to config file.}
}
\value{
A list of DBIConnectors
}
\description{
Read connectors
}
