% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbiconf-package.R
\docType{package}
\name{dbiconf-package}
\alias{dbiconf}
\alias{dbiconf-package}
\title{dbiconf: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Shun Sambongi \email{shundev3@gmail.com}

}
\keyword{internal}
